requirements.txt
# Core dependencies
tweepy>=4.14.0
python-telegram-bot>=13.15
fastapi>=0.97.0
uvicorn>=0.22.0
pydantic>=2.0.0
SQLAlchemy>=2.0.0
asyncio>=3.4.3
aiohttp>=3.8.4

# Configuration and logging
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Database drivers
aiosqlite>=0.18.0  # For SQLite
# psycopg2-binary>=2.9.6  # Uncomment for PostgreSQL

# Utils
pytz>=2023.3
tenacity>=8.2.2  # For retries
aiofiles>=23.1.0
cryptography>=41.0.1 